requirements.txt
chardet==2.3.0
command-not-found==0.3
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-toolbelt==0.0.1
gunicorn==19.6.0
heroku==0.1.4
httpsig==1.1.2
Kivy==1.9.0
language-selector==0.1
nose==1.3.7
numpy==1.11.1
pandas==0.18.1
psycopg2==2.6.2
pycrypto==2.6.1
pycurl==7.43.0
pygobject==3.20.0
python-apt==1.1.0b1
python-dateutil==2.5.3
python-debian==0.1.27
python-dwapi==0.1.dev0
python-systemd==231
pytz==2016.6.1
requests==2.9.1
six==1.10.0
ssh-import-id==5.5
static3==0.7.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
